requirements.txt
langchain
langgraph
langchain-openai
langchain-ollama
langchain-chroma
langchain-community
pillow
ipython
psycopg
langgraph-checkpoint-postgres
chainlit
fastapi
uvicorn
beautifulsoup4
util